clean up for building toolbar
