+++ src/lib/app/RvCommon/CMakeLists.txt
@@ -55,6 +55,7 @@ SET(_sources
     RvWebManager.cpp
     RvTopViewToolBar.cpp
     RvBottomViewToolBar.cpp
+    RvBottomViewToolBarBuild.cpp
     RvSourceEditor.cpp
     GLSLSyntaxHighlighter.cpp
     RvProfileManager.cpp
