+++ src/lib/app/RvCommon/RvBottomViewToolBar.cpp
@@ -15,7 +15,6 @@
 #include <QtWidgets/QToolTip>
 #include <QAction>
 #include <QtWidgets/QWidgetAction>
-#include <QtWidgets/QFrame>
 #include <QtWidgets/QHBoxLayout>
 #include <QtWidgets/QWidget>
 #include <IPCore/Session.h>
@@ -71,24 +70,11 @@ namespace Rv
         setVisible(b);
     }
 
-    static QVariant enumCodeMap(int enumVal, QString codeVal)
-    {
-        QVariantMap m;
-
-        m["enum"] = enumVal;
-        m["code"] = codeVal;
-
-        return QVariant(m);
-    }
-
     void RvBottomViewToolBar::build()
     {
         if (m_audioMenu)
             return;
 
-        Options& opts = Options::sharedOptions();
-
-        //  Initialize QIcons
         m_playModeOnceIcon = QIcon(":/images/playmode_once_48x48.png");
         m_playModeLoopIcon = QIcon(":/images/playmode_loop_48x48.png");
         m_playModePingPongIcon = QIcon(":/images/playmode_pingpong_48x48.png");
@@ -104,216 +90,12 @@ namespace Rv
         constexpr int sideBoxPadding = 20;
         constexpr int sideBoxFixedWidth = 450;
 
-        auto makeBtn = [](QWidget* parent, QAction* action, const char* tbstyle, const char* tbsize = nullptr) {
-            QToolButton* btn = new QToolButton(parent);
-            btn->setDefaultAction(action);
-            btn->setProperty("tbstyle", QVariant(QString(tbstyle)));
-            if (tbsize)
-                btn->setProperty("tbsize", QVariant(QString(tbsize)));
-            btn->setToolButtonStyle(Qt::ToolButtonIconOnly);
-            return btn;
-        };
-
-        auto makeExpandingSpacer = [this]() {
-            QFrame* qf = new QFrame(this);
-            qf->setSizePolicy(QSizePolicy::Expanding, QSizePolicy::Minimum);
-            qf->setMinimumWidth(0);
-            qf->setFrameStyle(QFrame::Plain | QFrame::NoFrame);
-            qf->setStyleSheet("background-color: transparent");
-            return qf;
-        };
-
-        m_leftBox = new QWidget(this);
-        m_leftBox->setObjectName("leftBox");
-        QHBoxLayout* leftLayout = new QHBoxLayout(m_leftBox);
-        leftLayout->setContentsMargins(sideBoxPadding, 0, sideBoxPadding, 0);
-        leftLayout->setSpacing(0);
-        m_leftBox->setFixedWidth(sideBoxFixedWidth);
-
-        struct LeftActionDef
-        {
-            const char* iconPath;
-            const char* tooltip;
-            const char* tbstyle;
-            QAction** target;
-        };
-        const LeftActionDef leftDefs[] = {
-            {":/images/smanager.png",     "Toggle Session Manager",      "left",     &m_smAction},
-            {":/images/paint_48x48.png",  "Toggle Annotation tools",     "interior", &m_paintAction},
-            {":/images/about_48x48.png",  "Toggle Image Info",           "interior", &m_infoAction},
-            {":/images/ntwrk_48x48.png",  "Toggle RV Networking Dialog", "interior", &m_networkAction},
-            {":/images/timeline_mag.png", "Toggle Timeline Magnifier",   "interior", &m_timelineMagAction},
-            {":/images/timeline.png",     "Toggle Timeline",             "right",    &m_timelineAction},
-        };
-        for (const auto& def : leftDefs)
-        {
-            QAction* a = new QAction("", this);
-            a->setIcon(QIcon(def.iconPath));
-            a->setToolTip(def.tooltip);
-            *def.target = a;
-            leftLayout->addWidget(makeBtn(m_leftBox, a, def.tbstyle));
-        }
-
-        connect(m_smAction, SIGNAL(triggered(bool)), this, SLOT(smActionTriggered(bool)));
-        connect(m_paintAction, SIGNAL(triggered(bool)), this, SLOT(paintActionTriggered(bool)));
-        connect(m_infoAction, SIGNAL(triggered(bool)), this, SLOT(infoActionTriggered(bool)));
-        connect(m_timelineAction, SIGNAL(triggered(bool)), this, SLOT(timelineActionTriggered(bool)));
-        connect(m_timelineMagAction, SIGNAL(triggered(bool)), this, SLOT(timelineMagActionTriggered(bool)));
-        connect(m_networkAction, SIGNAL(triggered(bool)), this, SLOT(networkActionTriggered(bool)));
-
-        m_ghostAction = new QAction("", this);
-        m_ghostAction->setIcon(QIcon(":/images/ghost.png"));
-        m_ghostAction->setToolTip("Ghost");
-        m_ghostAction->setCheckable(true);
-        leftLayout->addWidget(makeBtn(m_leftBox, m_ghostAction, "left"));
-
-        m_holdAction = new QAction("", this);
-        m_holdAction->setIcon(QIcon(":/images/hold.png"));
-        m_holdAction->setToolTip("Hold");
-        m_holdAction->setCheckable(true);
-        leftLayout->addWidget(makeBtn(m_leftBox, m_holdAction, "right"));
-
-        connect(m_ghostAction, SIGNAL(triggered(bool)), this, SLOT(ghostTriggered(bool)));
-        connect(m_holdAction, SIGNAL(triggered(bool)), this, SLOT(holdTriggered(bool)));
-
-        leftLayout->addStretch();
-        addWidget(m_leftBox);
-
+        buildLeft(sideBoxPadding, sideBoxFixedWidth);
         addWidget(makeExpandingSpacer());
-
-        m_centerBox = new QWidget(this);
-        m_centerBox->setObjectName("centerBox");
-        QHBoxLayout* centerLayout = new QHBoxLayout(m_centerBox);
-        centerLayout->setContentsMargins(0, 0, 0, 0);
-        centerLayout->setSpacing(0);
-
-        m_backStepAction = new QAction("", this);
-        m_backStepAction->setIcon(QIcon(":/images/control_bstep.png"));
-        m_backStepAction->setToolTip("Step back one frame");
-        QToolButton* backStepBtn = makeBtn(m_centerBox, m_backStepAction, "left");
-        backStepBtn->setObjectName("backStepButton");
-        centerLayout->addWidget(backStepBtn);
-
-        m_forwardStepAction = new QAction("", this);
-        m_forwardStepAction->setIcon(QIcon(":/images/control_fstep.png"));
-        m_forwardStepAction->setToolTip("Step forward one frame");
-        QToolButton* fwdStepBtn = makeBtn(m_centerBox, m_forwardStepAction, "right");
-        fwdStepBtn->setObjectName("forwardStepButton");
-        centerLayout->addWidget(fwdStepBtn);
-
-        m_backwardPlayAction = new QAction("", this);
-        m_backwardPlayAction->setIcon(QIcon(":/images/control_bplay.png"));
-        m_backwardPlayAction->setToolTip("Play backwards");
-        centerLayout->addWidget(makeBtn(m_centerBox, m_backwardPlayAction, "left", "double"));
-
-        m_forwardPlayAction = new QAction("", this);
-        m_forwardPlayAction->setIcon(QIcon(":/images/control_play.png"));
-        m_forwardPlayAction->setToolTip("Play forwards");
-        centerLayout->addWidget(makeBtn(m_centerBox, m_forwardPlayAction, "right", "double"));
-
-        m_backMarkAction = new QAction("", this);
-        m_backMarkAction->setIcon(QIcon(":/images/control_bmark.png"));
-        m_backMarkAction->setToolTip("Skip to start of sequence");
-        QToolButton* backMarkBtn = makeBtn(m_centerBox, m_backMarkAction, "left");
-        backMarkBtn->setObjectName("firstFrameButton");
-        centerLayout->addWidget(backMarkBtn);
-
-        m_forwardMarkAction = new QAction("", this);
-        m_forwardMarkAction->setIcon(QIcon(":/images/control_fmark.png"));
-        m_forwardMarkAction->setToolTip("Skip to end of sequence");
-        QToolButton* fwdMarkBtn = makeBtn(m_centerBox, m_forwardMarkAction, "right");
-        fwdMarkBtn->setObjectName("lastFrameButton");
-        centerLayout->addWidget(fwdMarkBtn);
-
-        connect(m_backStepAction, SIGNAL(triggered()), this, SLOT(backStepTriggered()));
-        connect(m_forwardStepAction, SIGNAL(triggered()), this, SLOT(forwardStepTriggered()));
-        connect(m_backwardPlayAction, SIGNAL(triggered()), this, SLOT(backPlayTriggered()));
-        connect(m_forwardPlayAction, SIGNAL(triggered()), this, SLOT(forwardPlayTriggered()));
-        connect(m_backMarkAction, SIGNAL(triggered()), this, SLOT(backMarkTriggered()));
-        connect(m_forwardMarkAction, SIGNAL(triggered()), this, SLOT(forwardMarkTriggered()));
-
-        addWidget(m_centerBox);
-
+        buildCenter();
         addWidget(makeExpandingSpacer());
+        buildRight(sideBoxPadding, sideBoxFixedWidth);
 
-        m_rightBox = new QWidget(this);
-        m_rightBox->setObjectName("rightBox");
-        QHBoxLayout* rightLayout = new QHBoxLayout(m_rightBox);
-        rightLayout->setContentsMargins(sideBoxPadding, 0, sideBoxPadding, 0);
-        rightLayout->setSpacing(0);
-        m_rightBox->setFixedWidth(sideBoxFixedWidth);
-
-        rightLayout->addStretch();
-
-        m_playModeAction = new QAction("", this);
-        m_playModeAction->setToolTip(QString::fromUtf8(playModeDefaultTooltip.data(), playModeDefaultTooltip.size()));
-        switch (static_cast<Session::PlayMode>(opts.loopMode))
-        {
-        case Session::PlayOnce:
-            m_playModeAction->setIcon(m_playModeOnceIcon);
-            break;
-        case Session::PlayPingPong:
-            m_playModeAction->setIcon(m_playModePingPongIcon);
-            break;
-        default:
-        case Session::PlayLoop:
-            m_playModeAction->setIcon(m_playModeLoopIcon);
-            break;
-        }
-
-        QToolButton* playModeBtn = makeBtn(m_rightBox, m_playModeAction, "left_menu");
-        playModeBtn->setObjectName("playModeButton");
-        playModeBtn->setPopupMode(QToolButton::InstantPopup);
-
-        m_playModeMenu = new QMenu(playModeBtn);
-        m_playModeMenu->addAction("Playback")->setDisabled(true);
-        m_playModeOnceAction = m_playModeMenu->addAction("  Once");
-        m_playModeOnceAction->setData(enumCodeMap(Session::PlayOnce, "commands.setPlayMode(PlayOnce)"));
-        m_playModeLoopAction = m_playModeMenu->addAction("  Loop");
-        m_playModeLoopAction->setData(enumCodeMap(Session::PlayLoop, "commands.setPlayMode(PlayLoop)"));
-        m_playModePingPongAction = m_playModeMenu->addAction("  PingPong");
-        m_playModePingPongAction->setData(enumCodeMap(Session::PlayPingPong, "commands.setPlayMode(PlayPingPong)"));
-        playModeBtn->setMenu(m_playModeMenu);
-        rightLayout->addWidget(playModeBtn);
-
-        m_playModeMenu->installEventFilter(this);
-        connect(m_playModeMenu, SIGNAL(triggered(QAction*)), this, SLOT(playModeMenuTriggered(QAction*)));
-        connect(m_playModeMenu, SIGNAL(aboutToShow()), this, SLOT(playModeMenuUpdate()));
-
-        int volumeLevel = (int)(100.0f * IPCore::SoundTrackIPNode::defaultVolume);
-        m_audioAction = new QAction("", this);
-        m_audioAction->setToolTip("Audio control");
-        setVolumeLevel<QAction>(*m_audioAction, volumeLevel);
-
-        QToolButton* audioBtn = makeBtn(m_rightBox, m_audioAction, "right_menu");
-        audioBtn->setPopupMode(QToolButton::InstantPopup);
-
-        m_audioMenu = new QMenu(audioBtn);
-        m_audioMenu->setProperty("menuStyle", QVariant(QString("slim")));
-        QAction* audioLabel = m_audioMenu->addAction("Volume");
-        audioLabel->setDisabled(true);
-        QWidgetAction* wa = new QWidgetAction(audioBtn);
-        setVolumeLevel<QWidgetAction>(*wa, volumeLevel);
-        m_audioSlider = new QSlider(audioBtn);
-        m_audioSlider->setProperty("sliderStyle", QVariant(QString("menu")));
-        m_audioSlider->setTickInterval(10);
-        wa->setDefaultWidget(m_audioSlider);
-        m_audioMenu->addAction(wa);
-        m_audioMenu->addSeparator();
-        m_muteAction = m_audioMenu->addAction("Mute");
-        m_muteAction->setIcon(QIcon(":/images/mute_32x32.png"));
-        m_muteAction->setCheckable(true);
-        audioBtn->setMenu(m_audioMenu);
-        rightLayout->addWidget(audioBtn);
-
-        connect(m_muteAction, SIGNAL(triggered(bool)), this, SLOT(audioMuteTriggered(bool)));
-        connect(m_audioMenu, SIGNAL(aboutToShow()), this, SLOT(audioMenuTriggered()));
-        connect(m_audioSlider, SIGNAL(valueChanged(int)), this, SLOT(audioSliderChanged(int)));
-        connect(m_audioSlider, SIGNAL(sliderReleased()), this, SLOT(audioSliderReleased()));
-
-        addWidget(m_rightBox);
-
-        // Map toolbar actions to their corresponding event categories
         m_actionCategoryMappings = {{
             {m_smAction, IPCore::EventCategories::sessionmanagerCategory, m_smAction->toolTip()},
             {m_paintAction, IPCore::EventCategories::annotateCategory, m_paintAction->toolTip()},
@@ -324,7 +106,6 @@ namespace Rv
             {m_backwardPlayAction, IPCore::EventCategories::backwardplayCategory, m_backwardPlayAction->toolTip()},
             {m_forwardPlayAction, IPCore::EventCategories::playcontrolCategory, m_forwardPlayAction->toolTip()},
             {m_playModeAction, IPCore::EventCategories::playcontrolCategory, m_playModeAction->toolTip()},
-
             {m_backMarkAction, IPCore::EventCategories::playcontrolCategory, m_backMarkAction->toolTip()},
             {m_forwardMarkAction, IPCore::EventCategories::playcontrolCategory, m_forwardMarkAction->toolTip()},
         }};
